# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  ent.Classdate:
    properties:
      DAY:
        description: DAY holds the value of the "DAY" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ClassdateEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ClassdateQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ClassdateEdges:
    properties:
      courseclasses:
        description: Courseclasses holds the value of the courseclasses edge.
        items:
          $ref: '#/definitions/ent.Courseclass'
        type: array
    type: object
  ent.Classroom:
    properties:
      ROOM:
        description: ROOM holds the value of the "ROOM" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ClassroomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ClassroomQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ClassroomEdges:
    properties:
      courseclasses:
        description: Courseclasses holds the value of the courseclasses edge.
        items:
          $ref: '#/definitions/ent.Courseclass'
        type: array
    type: object
  ent.Classtime:
    properties:
      TIME:
        description: TIME holds the value of the "TIME" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ClasstimeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ClasstimeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ClasstimeEdges:
    properties:
      courseclasses:
        description: Courseclasses holds the value of the courseclasses edge.
        items:
          $ref: '#/definitions/ent.Courseclass'
        type: array
    type: object
  ent.Course:
    properties:
      Course_name:
        description: CourseName holds the value of the "Course_name" field.
        type: string
      Course_year:
        description: CourseYear holds the value of the "Course_year" field.
        type: string
      Teacher_id:
        description: TeacherID holds the value of the "Teacher_id" field.
        type: string
      degree_id:
        type: integer
      edges:
        $ref: '#/definitions/ent.CourseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CourseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      subject_id:
        type: integer
    type: object
  ent.CourseEdges:
    properties:
      degreeID:
        $ref: '#/definitions/ent.Degree'
        description: DegreeID holds the value of the Degree_id edge.
        type: object
      departmentID:
        $ref: '#/definitions/ent.Department'
        description: DepartmentID holds the value of the Department_id edge.
        type: object
      subjectID:
        $ref: '#/definitions/ent.Subject'
        description: SubjectID holds the value of the Subject_id edge.
        type: object
    type: object
  ent.Courseclass:
    properties:
      edges:
        $ref: '#/definitions/ent.CourseclassEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CourseclassQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      instructorInfo_id:
        type: integer
      tablecode:
        description: Tablecode holds the value of the "tablecode" field.
        type: string
    type: object
  ent.CourseclassEdges:
    properties:
      classdate:
        $ref: '#/definitions/ent.Classdate'
        description: Classdate holds the value of the classdate edge.
        type: object
      classroom:
        $ref: '#/definitions/ent.Classroom'
        description: Classroom holds the value of the classroom edge.
        type: object
      classtime:
        $ref: '#/definitions/ent.Classtime'
        description: Classtime holds the value of the classtime edge.
        type: object
      instructorInfo:
        $ref: '#/definitions/ent.InstructorInfo'
        description: InstructorInfo holds the value of the instructorInfo edge.
        type: object
      subject:
        $ref: '#/definitions/ent.Subject'
        description: Subject holds the value of the subject edge.
        type: object
    type: object
  ent.Degree:
    properties:
      Degree_name:
        description: DegreeName holds the value of the "Degree_name" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DegreeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DegreeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.DegreeEdges:
    properties:
      degree:
        description: Degree holds the value of the degree edge.
        items:
          $ref: '#/definitions/ent.Course'
        type: array
      subjectsOffered:
        description: SubjectsOffered holds the value of the SubjectsOffered edge.
        items:
          $ref: '#/definitions/ent.SubjectsOffered'
        type: array
    type: object
  ent.Department:
    properties:
      DEPARTMENT:
        description: DEPARTMENT holds the value of the "DEPARTMENT" field.
        type: string
      FACULTY:
        description: FACULTY holds the value of the "FACULTY" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DepartmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DepartmentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.DepartmentEdges:
    properties:
      department:
        description: Department holds the value of the department edge.
        items:
          $ref: '#/definitions/ent.Course'
        type: array
      instructorinfos:
        description: Instructorinfos holds the value of the instructorinfos edge.
        items:
          $ref: '#/definitions/ent.InstructorInfo'
        type: array
    type: object
  ent.InstructorInfo:
    properties:
      EMAIL:
        description: EMAIL holds the value of the "EMAIL" field.
        type: string
      NAME:
        description: NAME holds the value of the "NAME" field.
        type: string
      PASSWORD:
        description: PASSWORD holds the value of the "PASSWORD" field.
        type: string
      PHONENUMBER:
        description: PHONENUMBER holds the value of the "PHONENUMBER" field.
        type: string
      edges:
        $ref: '#/definitions/ent.InstructorInfoEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the InstructorInfoQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.InstructorInfoEdges:
    properties:
      courseclasses:
        description: Courseclasses holds the value of the courseclasses edge.
        items:
          $ref: '#/definitions/ent.Courseclass'
        type: array
      department:
        $ref: '#/definitions/ent.Department'
        description: Department holds the value of the department edge.
        type: object
      instructorroom:
        $ref: '#/definitions/ent.InstructorRoom'
        description: Instructorroom holds the value of the instructorroom edge.
        type: object
      title:
        $ref: '#/definitions/ent.Title'
        description: Title holds the value of the title edge.
        type: object
    type: object
  ent.InstructorRoom:
    properties:
      BUILDING:
        description: BUILDING holds the value of the "BUILDING" field.
        type: string
      ROOM:
        description: ROOM holds the value of the "ROOM" field.
        type: string
      edges:
        $ref: '#/definitions/ent.InstructorRoomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the InstructorRoomQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.InstructorRoomEdges:
    properties:
      instructorinfos:
        description: Instructorinfos holds the value of the instructorinfos edge.
        items:
          $ref: '#/definitions/ent.InstructorInfo'
        type: array
    type: object
  ent.Subject:
    properties:
      Subject_name:
        description: SubjectName holds the value of the "Subject_name" field.
        type: string
      edges:
        $ref: '#/definitions/ent.SubjectEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SubjectQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.SubjectEdges:
    properties:
      courseclasses:
        description: Courseclasses holds the value of the courseclasses edge.
        items:
          $ref: '#/definitions/ent.Courseclass'
        type: array
      subject:
        description: Subject holds the value of the subject edge.
        items:
          $ref: '#/definitions/ent.Course'
        type: array
      subjectsOffered:
        description: SubjectsOffered holds the value of the SubjectsOffered edge.
        items:
          $ref: '#/definitions/ent.SubjectsOffered'
        type: array
    type: object
  ent.SubjectsOffered:
    properties:
      AMOUNT:
        description: AMOUNT holds the value of the "AMOUNT" field.
        type: string
      Remain:
        description: Remain holds the value of the "Remain" field.
        type: string
      STATUS:
        description: STATUS holds the value of the "STATUS" field.
        type: string
      degree_id:
        type: integer
      edges:
        $ref: '#/definitions/ent.SubjectsOfferedEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SubjectsOfferedQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      subject_id:
        type: integer
    type: object
  ent.SubjectsOfferedEdges:
    properties:
      degree:
        $ref: '#/definitions/ent.Degree'
        description: Degree holds the value of the Degree edge.
        type: object
      subject:
        $ref: '#/definitions/ent.Subject'
        description: Subject holds the value of the Subject edge.
        type: object
      term:
        $ref: '#/definitions/ent.Term'
        description: Term holds the value of the Term edge.
        type: object
      year:
        $ref: '#/definitions/ent.Year'
        description: Year holds the value of the Year edge.
        type: object
    type: object
  ent.Term:
    properties:
      TERM:
        description: TERM holds the value of the "TERM" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.TermEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TermQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.TermEdges:
    properties:
      subjectsOffered:
        description: SubjectsOffered holds the value of the SubjectsOffered edge.
        items:
          $ref: '#/definitions/ent.SubjectsOffered'
        type: array
    type: object
  ent.Title:
    properties:
      TITLE:
        description: TITLE holds the value of the "TITLE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.TitleEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TitleQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.TitleEdges:
    properties:
      instructorinfos:
        description: Instructorinfos holds the value of the instructorinfos edge.
        items:
          $ref: '#/definitions/ent.InstructorInfo'
        type: array
    type: object
  ent.Year:
    properties:
      YEAR:
        description: YEAR holds the value of the "YEAR" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.YearEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the YearQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.YearEdges:
    properties:
      subjectsOffered:
        description: SubjectsOffered holds the value of the SubjectsOffered edge.
        items:
          $ref: '#/definitions/ent.SubjectsOffered'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API
  version: "1.0"
paths:
...
  /instructorinfos:
    get:
      description: list instructorinfo entities
      operationId: list-instructorinfo
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.InstructorInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List instructorinfo entities
    post:
      consumes:
      - application/json
      description: Create instructorinfo
      operationId: create-instructorinfo
      parameters:
      - description: InstructorInfo entity
        in: body
        name: instructorinfo
        required: true
        schema:
          $ref: '#/definitions/ent.InstructorInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.InstructorInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create instructorinfo
  /instructorinfos/{id}:
    delete:
      description: get instructorinfo by ID
      operationId: delete-instructorinfo
      parameters:
      - description: InstructorInfo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a instructorinfo entity by ID
    get:
      description: get instructorinfo by ID
      operationId: get-instructorinfo
      parameters:
      - description: InstructorInfo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.InstructorInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a instructorinfo entity by ID
    put:
      consumes:
      - application/json
      description: update instructorinfo by ID
      operationId: update-instructorinfo
      parameters:
      - description: InstructorInfo ID
        in: path
        name: id
        required: true
        type: integer
      - description: InstructorInfo entity
        in: body
        name: instructorinfo
        required: true
        schema:
          $ref: '#/definitions/ent.InstructorInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.InstructorInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update a instructorinfo entity by ID
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    authorizationUrl: ""
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    authorizationUrl: ""
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"